feat: finish implementation
