+++ mlflow/genai/judges/tools/constants.py
@@ -19,5 +19,5 @@ class ToolNames:
     LIST_SPANS = "list_spans"
     SEARCH_TRACE_REGEX = "search_trace_regex"
     GET_SPAN_PERFORMANCE_AND_TIMING_REPORT = "get_span_performance_and_timing_report"
-    GET_HISTORICAL_TRACES = "get_historical_traces"
+    GET_TRACES_IN_SESSION = "get_traces_in_session"
     SEARCH_TRACES = "search_traces"
