+++ mlflow/genai/judges/tools/registry.py
@@ -127,13 +127,13 @@ def list_judge_tools() -> list[JudgeTool]:
 
 # NB: Tool imports are at the bottom to avoid circular dependencies and ensure
 # the registry is fully defined before tools attempt to register themselves.
-from mlflow.genai.judges.tools.get_historical_traces import GetHistoricalTracesTool
 from mlflow.genai.judges.tools.get_root_span import GetRootSpanTool
 from mlflow.genai.judges.tools.get_span import GetSpanTool
 from mlflow.genai.judges.tools.get_span_performance_and_timing_report import (
     GetSpanPerformanceAndTimingReportTool,
 )
 from mlflow.genai.judges.tools.get_trace_info import GetTraceInfoTool
+from mlflow.genai.judges.tools.get_traces_in_session import GetTracesInSession
 from mlflow.genai.judges.tools.list_spans import ListSpansTool
 from mlflow.genai.judges.tools.search_trace_regex import SearchTraceRegexTool
 
@@ -143,4 +143,4 @@ _judge_tool_registry.register(GetSpanTool())
 _judge_tool_registry.register(ListSpansTool())
 _judge_tool_registry.register(SearchTraceRegexTool())
 _judge_tool_registry.register(GetSpanPerformanceAndTimingReportTool())
-_judge_tool_registry.register(GetHistoricalTracesTool())
+_judge_tool_registry.register(GetTracesInSession())
