+++ tests/genai/judges/test_get_traces_in_session_tool.py
@@ -1,0 +1,167 @@
+"""
+Tests for GetTracesInSession.
+
+This module tests the functionality of the GetTracesInSession class,
+including successful trace retrieval, error handling, and parameter validation.
+"""
+
+from unittest.mock import MagicMock, patch
+
+import pytest
+
+from mlflow.entities.trace import Trace
+from mlflow.entities.trace_info import TraceInfo as MlflowTraceInfo
+from mlflow.entities.trace_location import TraceLocation
+from mlflow.entities.trace_state import TraceState
+from mlflow.exceptions import MlflowException
+from mlflow.genai.judges.tools.get_traces_in_session import GetTracesInSession
+from mlflow.genai.judges.tools.types import TraceInfo
+from mlflow.protos.databricks_pb2 import INVALID_PARAMETER_VALUE, ErrorCode
+from mlflow.types.llm import ToolDefinition
+
+
+def test_get_traces_in_session_tool_name():
+    """Test that the tool returns the correct name."""
+    tool = GetTracesInSession()
+    assert tool.name == "get_traces_in_session"
+
+
+def test_get_traces_in_session_tool_get_definition():
+    """Test that the tool returns a valid tool definition."""
+    tool = GetTracesInSession()
+    definition = tool.get_definition()
+
+    assert isinstance(definition, ToolDefinition)
+    assert definition.function.name == "get_traces_in_session"
+    assert "session" in definition.function.description.lower()
+    assert "multi-turn evaluation" in definition.function.description.lower()
+    assert definition.function.parameters.type == "object"
+    assert len(definition.function.parameters.required) == 0
+    assert definition.type == "function"
+
+    params = definition.function.parameters.properties
+    assert "max_results" in params
+    assert "order_by" in params
+    assert params["max_results"].type == "integer"
+    assert params["order_by"].type == "array"
+
+
+def create_mock_trace(trace_id: str, experiment_id: str = "exp-123", session_id: str | None = None):
+    """Helper function to create mock trace objects."""
+    tags = {}
+    if session_id:
+        tags["session.id"] = session_id
+
+    trace_info = MlflowTraceInfo(
+        trace_id=trace_id,
+        trace_location=TraceLocation.from_experiment_id(experiment_id),
+        request_time=1234567890,
+        state=TraceState.OK,
+        execution_duration=250,
+        tags=tags,
+    )
+    return Trace(info=trace_info, data=None)
+
+
+@patch("mlflow.genai.judges.tools.get_traces_in_session.SearchTracesTool")
+def test_get_traces_in_session_tool_invoke_success(mock_search_tool_class):
+    """Test successful retrieval of traces in session."""
+    tool = GetTracesInSession()
+    current_trace = create_mock_trace("current-trace", "exp-123", "session-123")
+
+    mock_search_tool = MagicMock()
+    mock_result = [
+        TraceInfo(
+            trace_id="trace-1",
+            request_time=1000,
+            state=TraceState.OK,
+            request="What is machine learning?",
+            response="Machine learning is a subset of AI...",
+            execution_duration=100,
+            assessments=[],
+        ),
+        TraceInfo(
+            trace_id="trace-2",
+            request_time=2000,
+            state=TraceState.OK,
+            request="Can you give an example?",
+            response="Sure! A common example is...",
+            execution_duration=150,
+            assessments=[],
+        ),
+    ]
+    mock_search_tool.invoke.return_value = mock_result
+    mock_search_tool_class.return_value = mock_search_tool
+
+    result = tool.invoke(current_trace)
+
+    assert len(result) == 2
+    assert all(isinstance(ti, TraceInfo) for ti in result)
+    assert result[0].trace_id == "trace-1"
+    assert result[0].request == "What is machine learning?"
+    assert result[1].trace_id == "trace-2"
+
+    mock_search_tool.invoke.assert_called_once_with(
+        current_trace,
+        "tags.`session.id` = 'session-123' AND trace.timestamp < 1234567890",
+        ["timestamp ASC"],
+        20,
+    )
+
+
+@patch("mlflow.genai.judges.tools.get_traces_in_session.SearchTracesTool")
+def test_get_traces_in_session_tool_invoke_custom_parameters(mock_search_tool_class):
+    """Test tool invocation with custom parameters."""
+    tool = GetTracesInSession()
+    current_trace = create_mock_trace("current-trace", "exp-123", "session-456")
+
+    mock_search_tool = MagicMock()
+    mock_search_tool.invoke.return_value = []
+    mock_search_tool_class.return_value = mock_search_tool
+
+    tool.invoke(current_trace, max_results=50, order_by=["timestamp DESC"])
+
+    mock_search_tool.invoke.assert_called_once_with(
+        current_trace,
+        "tags.`session.id` = 'session-456' AND trace.timestamp < 1234567890",
+        ["timestamp DESC"],
+        50,
+    )
+
+
+def test_get_traces_in_session_tool_invoke_no_session_id():
+    """Test that tool raises error when no session.id is found in trace tags."""
+    tool = GetTracesInSession()
+    current_trace = create_mock_trace("current-trace", "exp-123", session_id=None)
+
+    with pytest.raises(MlflowException, match="No session.id found in trace tags") as exc_info:
+        tool.invoke(current_trace)
+
+    assert exc_info.value.error_code == ErrorCode.Name(INVALID_PARAMETER_VALUE)
+
+
+def test_get_traces_in_session_tool_invoke_invalid_session_id():
+    """Test error when session.id has invalid format."""
+    tool = GetTracesInSession()
+    current_trace = create_mock_trace("current-trace", "exp-123", "session@123!invalid")
+
+    with pytest.raises(MlflowException, match="Invalid session ID format") as exc_info:
+        tool.invoke(current_trace)
+
+    assert exc_info.value.error_code == ErrorCode.Name(INVALID_PARAMETER_VALUE)
+
+
+@patch("mlflow.genai.judges.tools.get_traces_in_session.SearchTracesTool")
+def test_get_traces_in_session_tool_invoke_empty_result(mock_search_tool_class):
+    """Test tool behavior when no traces are found."""
+    tool = GetTracesInSession()
+    current_trace = create_mock_trace("current-trace", "exp-123", "session-123")
+
+    mock_search_tool = MagicMock()
+    mock_search_tool.invoke.return_value = []
+    mock_search_tool_class.return_value = mock_search_tool
+
+    result = tool.invoke(current_trace)
+
+    assert result == []
+    assert len(result) == 0
